flask-backend/requirement.txt
Flask-Cor==3.0.10
gunicorn==20.1.0
Jinja2==3.0.1	
MarkupSafe==2.0.1	
PySocks==1.7.1	
Werkzeug==2.0.1	
argparse==1.4.0	
certifi==2021.5.30	
cffi==1.14.5	
chardet==4.0.0	
click==8.0.1	
colorama==0.4.4	
cors==1.0.1	
filelock==3.0.12	
future==0.18.2	
gevent==21.1.2	
greenlet==1.1.0	
idna==2.10	
itsdangerous==2.0.1	
pip==21.1.2	
pycparser==2.20	
requests==2.25.1	
requests-file==1.5.1	
setuptools	==57.0.0	
six==1.16.0	
tldextract==3.1.0	
urllib3==1.26.5	
zope.event==4.5.0	
zope.interface==5.4.0	